change optimizer base for LarsMomentumOptimizer
