+++ python/paddle/incubate/optimizer/__init__.py
@@ -14,6 +14,7 @@
 
 from .lookahead import LookAhead  # noqa: F401
 from .modelaverage import ModelAverage  # noqa: F401
+from .lars_momentum import LarsMomentumOptimizer  # noqa: F401
 from .distributed_fused_lamb import DistributedFusedLamb  # noqa: F401
 from .lbfgs import LBFGS  # noqa: F401
 from . import functional  # noqa: F401
