+++ test/legacy_test/test_imperative_optimizer.py
@@ -28,7 +28,6 @@ from paddle.fluid.optimizer import (
     DpsgdOptimizer,
     ExponentialMovingAverage,
     FtrlOptimizer,
-    LarsMomentumOptimizer,
     LookaheadOptimizer,
     ModelAverage,
     MomentumOptimizer,
@@ -568,21 +567,6 @@ class TestImperativeMomentumOptimizer(TestImperativeOptimizerBase):
         return optimizer
 
     def test_momentum(self):
-        self._check_mlp()
-
-
-class TestImperativeLarsMomentumOptimizer(TestImperativeOptimizerBase):
-    def get_optimizer_dygraph(self, parameter_list):
-        optimizer = LarsMomentumOptimizer(
-            learning_rate=0.001, momentum=0.9, parameter_list=parameter_list
-        )
-        return optimizer
-
-    def get_optimizer(self):
-        optimizer = LarsMomentumOptimizer(learning_rate=0.001, momentum=0.9)
-        return optimizer
-
-    def test_larsmomentum(self):
         self._check_mlp()
